WIP: Tidy-up and re-jig before I go any further.
