+++ src/components/error-icon/template.vue
@@ -3,5 +3,5 @@
 </template>
 
 <script>
-export default {};
+    export default {};
 </script>
