+++ src/components/spinner/index.js
@@ -9,5 +9,4 @@ import Vue from 'vue';
 import spinner from './template.vue';
 
 Vue.component('spinner', spinner);
-
 export default spinner;
