+++ src/index.js
@@ -1,14 +1,5 @@
 /**
- * External
- */
-window.SimpleMDE = require('simplemde');
-
-/**
  * Internal
  */
+import './instances';
 import './sass/style.scss';
-
-/**
- * Loads all instances.
- */
-require('./instances/index');
