+++ src/instances/login/login.vue
@@ -7,21 +7,21 @@
         <login-form />
 
         <footer class="sign-off">
-            <a href="https://github.com/byrobots/writedown">WriteDown</a>,
+            <a href="https://github.com/by-robots/writedown">WriteDown</a>,
             a project <a href="https://by-robots.com">By Robots</a>.
         </footer>
     </div>
 </template>
 
 <script>
-/**
- * Internal.
- */
-import loginForm from '../../components/login-form';
-import template from '../../mixins/template.js';
+    /**
+     * Internal.
+     */
+    import loginForm from '../../components/login-form';
+    import template from '../../mixins/template.js';
 
-export default {
-    components: {loginForm},
-    mixins: [template],
-};
+    export default {
+        components: {loginForm},
+        mixins: [template],
+    };
 </script>
