+++ src/instances/posts/create/index.js
@@ -7,7 +7,7 @@ import Vue from 'vue';
  * Internal
  */
 import postCreate from './post-create.vue';
-import instance from '../../mixins/instance.js';
+import instance from '../../../mixins/instance.js';
 
 if (document.getElementById('post-create')) {
     new Vue({
