+++ src/instances/posts/create/post-create.vue
@@ -1,0 +1,22 @@
+<template>
+    <section>
+        <header>
+            <h1 class="pagetitle" v-text="pagetitle"/>
+        </header>
+
+        <post-form action="/admin/posts" />
+    </section>
+</template>
+
+<script>
+    /**
+     * Internal
+     */
+    import template from '../../../mixins/template.js';
+    import postForm from '../../../components/post-form';
+
+    export default {
+        components: {postForm},
+        mixins: [template],
+    };
+</script>
