+++ src/instances/posts/edit/index.js
@@ -7,7 +7,7 @@ import Vue from 'vue';
  * Internal
  */
 import postEdit from './post-edit.vue';
-import instance from '../../mixins/instance.js';
+import instance from '../../../mixins/instance.js';
 
 if (document.getElementById('post-edit')) {
     new Vue({
