+++ src/mixins/instance.js
@@ -8,7 +8,10 @@ import store from '../store';
  */
 export default {
     beforeMount () {
+        // The generated CSRF token to use with requests.
         store.commit('csrf', this.$el.attributes['data-csrf'].value);
+
+        // The page's title.
         store.commit('pagetitle', this.$el.attributes['data-pagetitle'].value);
     },
     store: store,
